Trying out calling the use cases services instead.
It's simpler to name things when you can append a suffix. In some cases
there just isn't anything in the domain that captures what the object
does, or the name is taken for a different purpose.

Another good reason to call them services is familarity. Everyone knows
what a service layer is.